+++ test/classes/volume.js
@@ -58,6 +58,20 @@
             readOnly: true
         });
 
+        let variableGlobalVolume = new Volume(new Layer(new Application('test', {
+            directories: {
+                test: {
+                    path: '/test/variable',
+                    description: 'Testing vartiable replacements!',
+                    shared: false
+                }
+            }
+        }), 'test', {}), {
+            host: '${test_global}/hello',
+            container: '/mnt/readonly',
+            readOnly: true
+        });
+
         it('should create a read only Volume', function () {
             expect(readOnlyVolume instanceof Volume).to.equal(true);
             expect(readOnlyVolume.readOnly).to.equal(true);
@@ -82,6 +96,10 @@
 
             it('should return the path to the volume on the host with variables', function () {
                 expect(variableVolume.host.indexOf('/test/variable/hello') > -1).to.equal(true);
+            });
+
+            it('should return the path to the volume on the host with variables', function () {
+                expect(variableGlobalVolume.host.indexOf('/test/hello') > -1).to.equal(true);
             });
         });